pyproject.toml
[tool.poetry]
name = "cinereserve"
version = "0.1.0"
description = "CineReserve API - Sistema de reservas do Cinépolis Natal"
authors = ["CineReserve Team <team@cinereserve.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
psycopg2-binary = "^2.9"
redis = "^5.0"
django-redis = "^5.4"
celery = "^5.3"
drf-spectacular = "^0.27"
python-decouple = "^3.8"
gunicorn = "^21.2"
django-filter = "^23.5"
Pillow = "^10.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.8"
pytest-cov = "^4.1"
factory-boy = "^3.3"
faker = "^23.0"
black = "^24.0"
flake8 = "^7.0"
isort = "^5.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 120
